refactor(nemotron_speech_streaming): specialize family implementation
Consolidate nemotron_speech_streaming model construction and weight mapping under its owned package. Remove unused family-local duplicates while preserving the existing plugin API.
