+++ python/tensorrt_model_connect/families/nemotron_speech_streaming/MODEL.toml
@@ -1,7 +1,7 @@
 id = "nemotron_speech_streaming"
 plugin = "nemotron_speech_streaming"
 module = "plugin"
-nemo_archive_adapter = "nemo_archive.py|resolve_nemo_archive"
+nemo_archive_adapter = "weights/nemo_archive.py|resolve_nemo_archive"
 nemo_model_type = "nemotron_speech_streaming"
 nemo_target_patterns = [
   "EncDecRNNT",
