--- python/tensorrt_model_connect/families/nemotron_speech_streaming/utils.py
@@ -1,83 +1,0 @@
-"""Model-agnostic helpers for TensorRT engine builders."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass
-
-import numpy as np
-from tensorrt_model_connect import trt_compat
-
-from . import graph_ops
-
-
-trt = trt_compat.get_trt()
-
-
-@dataclass(frozen=True)
-class BuilderContext:
-    """TensorRT objects shared by engine builders."""
-
-    logger: trt.Logger
-    builder: trt.Builder
-    network: trt.INetworkDefinition
-    config: trt.IBuilderConfig
-
-
-def create_builder_context(
-    *,
-    verbose: bool,
-    workspace_bytes: int,
-    strongly_typed: bool = True,
-    disable_tf32: bool = False,
-) -> BuilderContext:
-    """Create a TensorRT builder, network, and config with common defaults."""
-    logger = trt.Logger(trt.Logger.VERBOSE if verbose else trt.Logger.WARNING)
-    builder = trt.Builder(logger)
-    flags = 0
-    if strongly_typed:
-        flags |= 1 << int(trt.NetworkDefinitionCreationFlag.STRONGLY_TYPED)
-    network = builder.create_network(flags)
-    config = builder.create_builder_config()
-    config.set_memory_pool_limit(trt.MemoryPoolType.WORKSPACE, workspace_bytes)
-    if disable_tf32:
-        config.clear_flag(trt.BuilderFlag.TF32)
-    return BuilderContext(
-        logger=logger,
-        builder=builder,
-        network=network,
-        config=config,
-    )
-
-
-def const_in_work_dtype(
-    network: trt.INetworkDefinition,
-    shape: tuple,
-    values: np.ndarray,
-    work_np_dtype: np.dtype,
-    work_trt_dtype: trt.DataType,
-) -> trt.ITensor:
-    """Create a constant in storage dtype and cast it to runtime dtype."""
-    const = graph_ops.add_constant(network, shape, values, dtype=work_np_dtype)
-    if const.dtype != work_trt_dtype:
-        const = network.add_cast(const, work_trt_dtype).get_output(0)
-    return const
-
-
-def norm_multi(
-    network: trt.INetworkDefinition,
-    inp: trt.ITensor,
-    hidden: int,
-    gamma: np.ndarray,
-    beta: np.ndarray | None,
-    eps_tensor: trt.ITensor,
-    norm_type: str,
-    dtype: np.dtype,
-) -> trt.ITensor:
-    """Apply LayerNorm or RMSNorm from the same call site."""
-    if norm_type == "layernorm":
-        if beta is None:
-            beta = np.zeros(hidden, dtype=np.float32)
-        return graph_ops.add_layer_norm(
-            network, inp, hidden, gamma, beta, eps_tensor, dtype=dtype)
-    return graph_ops.add_rms_norm(
-        network, inp, hidden, gamma, eps_tensor, dtype=dtype)
